+++ tests/testthat/helpers.R
@@ -1,0 +1,73 @@
+#Helper functions for testing
+
+expect_good_matchit <- function(m, expect_subclass = NULL, expect_distance = NULL,
+                                expect_match.matrix = NULL, ratio = NULL, replace = NULL) {
+
+  if (isTRUE(expect_subclass) && isTRUE(replace)) {
+    stop("`expect_subclass` and `replace` cannot both be TRUE")
+  }
+  if (isTRUE(expect_match.matrix) && is.null(ratio)) {
+    stop("`ratio` cannot be NULL when `expect_match.matrix` is TRUE")
+  }
+
+  expect_s3_class(m, "matchit")
+
+  #Related to subclass
+  if (!is.null(expect_subclass)) {
+    if (expect_subclass) {
+      expect_false(is.null(m$subclass))
+      expect_length(m$sunclass, n)
+      expect_true(is.factor(m$subclass))
+      expect_false(is.null(names(m$subclass)))
+    }
+    else {
+      expect_null(m$subclass)
+    }
+  }
+
+  #Related to match.matrix
+  if (!is.null(expect_match.matrix)) {
+    if (expect_match.matrix) {
+      expect_false(is.null(m$match.matrix))
+      expect_type(m$match.matrix, "matrix")
+      expect_true(is.character(m$match.matrix))
+      expect_equal(nrow(m$match.matrix), n)
+      expect_equal(ncol(m$match.matrix), ratio)
+      expect_false(is.null(rownames(m$match.matrix)))
+      expect_false(any(rownames(m$match.matrix) %in% m$match.matrix))
+
+      #Check no duplicates within each row
+      expect_true(all(apply(m$match.matrix, 1, function(i) anyDuplicated(i[!is.na(i)]) == 0)))
+
+      if (!is.null(replace)) {
+        if (replace) {
+          #May not be duplicates incidentially; make sure examples induce duplicates
+          expect_true(!isTRUE(all.equal(anyDuplicated(m$match.matrix), 0)))
+        }
+        else {
+          expect_equal(anyDuplicated(m$match.matrix), 0)
+        }
+      }
+    }
+    else {
+      expect_null(m$match.matrix)
+    }
+  }
+
+  #Related to distance
+  if (!is.null(distance)) {
+    if (expect_distance) {
+      expect_false(is.null(m$distance))
+      expect_length(m$distance, n)
+      expect_true(is.numeric(m$distance))
+      expect_false(anyNA(m$distance))
+      expect_null(dim(m$distance))
+      expect_false(is.null(names(m$distance)))
+    }
+    else {
+      expect_null(m$distance)
+    }
+  }
+
+  invisible(m)
+}